Range slider now represents wallet transactions instead of block chain
To make this possible, the TransactionIndexRole was switched over to a
TransactionOrderPosRole that describes the transactional position in the
wallet.

MAX_BLOCKS_PER_PAGE has been renamed to MAX_TRANSACTIONS_PER_TICK. This
represents the constant much better now that we are controlling the wallet
range instead.

For this change to function, transaction records shown in the table had
to be extended to also include the transactional position.