Create transaction.cpp for handling transactions and transaction data
In order to get this working, this commit also moves certain disk
operations and disk logic to disk.cpp.

These changes only move some of the functions and classes that should be
here. The rest of them are still in main.cpp/h. These will be taken care
of with subsequent changes.