Fix QT build to include UPNP support by default
This builds with UPNP enabled if nothing is specified. Despite the build
description and documentation saying otherwise, the prior solution did not
build with UPNP enabled by default.

As it is intended mainly for servers, The CLI version still builds without
UPNP support enabled.

This change also adds a note to the cross-compile script, providing
information on how to build libminiupnpc under MXE.