Completely remove ui_interface.h and noui.cpp
This also removes all the unnecessary signals that were defined for the
sole purpose of connecting the backend daemon to the QT frontend. As we
will no longer be doing that - this code was completely redundant.

This also required some changes to how error messages are reported - some
are now written directly to the debug.log, while some are sent directly
to stderr.

For the localization support and the _()" function, we now use the
localization.h module that we defined in the previous commit.