Add libxz-embedded source code to the daemon
This source code is compact and easy to use. So we will use this to
decompress the bootstrap archives compressed with xz. In order to make the
archives compatible with the embedded library, some minor arguments and
tweaks were also added to the make_bootstrap.sh script. The integrity
checksum was set to CRC32 plus a few other additional options were also
supplied in order to allow xz to play nicely with this library.