The writing of address indexes is now a threaded operation
When writing these entries to the database, instead of waiting around for
a read/write operation to commit, the daemon will now fire off a thread
and run these operations in a seperate thread - allowing us to directly
continue processing the next incoming block.

Together with the previous change, the speed at which the wallet is now
able to process a single block has increased by a factor of ~20.