Add SSL compatibility layer with support for OpenSSL 1.1.x
The daemon can now be compiled in either OpenSSL 1.0.x or OpenSSL 1.1.x.
To get around the API changes to the more recent version a compatibility
layer is used (opensslcompat.c/h).

This (to my knowledge) makes Swipp the only cryptocurrency wallet in
existence to fully support both maintenance forks of OpenSSL.